Created Follow Agent Behavior Still WIP, but currently works as is. New Agent Action, Script used to to signal a script updated Agent.
Behavior Tree Updates
Fixed issue with moving Advancing the queue forward.
When the last node in traversal caused an error, the traversal would try to advance the queue forward, but because it was the last child and not in the queue anymore it would cause the queue to empty, ending the BT run.
The fix was to keep a secondary array of the ordered queue and look up the node to advance past and push the queue up to that point, not past the point.
Added way to remove invalid tree shapes.
Added Change Actor Behavior Tree Command.
Added Property Name to Behavior Tree State.
Updated string locations to use static property name.
More validation in Behavior Node Status.

Created Set Agent Behavior Command

New Scripts
SetAgentBehavior - Takes in agent global id and set to passed in behavior script.
SlowDown - Helps with debugging stuff by slowing down the BT.
FindNewMoveLocationAroundOwner - Find a random location around the Owner of the Actor.
HasNoOwner - Check to make sure that the Actor has NO owner.
IsFarFromOwner - Dictates if the Actor is far from the owner.
ResetToRunFromPlayer - When the Actor has now owner this is what BT they will reset to.
RunFromPlayer - Add some comments with State that will be created.