Fixed Agent Movement after Map Created/Rebuild To fix this issue I made two changes, Re-Register Agents on Map Changes and added a Path Monitor that will for Agents for stale paths.
Re-Register Agents on Map Changes
This will listen for the MapChangedEvent and trigger an un-register then register of the Agent, this will fix the behavior tree if it gets stuck during the update if it was already created and used.

Monitor for Stale Agent Paths
A timer was created that will trigger an event that will check for all agents for stale paths.
Stale Path requirements are Agent LocationState.CanMove is true, LocationState.NextMoveRequest older than 15 seconds, and PathState.Path.Count greater than 0.

Removed Path from Serialization from Agent Entity
This should fix an issue with the serialization of the agent when requests are sent to Clients.

41959132-7C95-4C22-A87F-EE5618BCDF0B